requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.2
Pillow==8.1.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
